pyproject.toml
[project]
name = "go-fluent-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "openai>=2.24.0",
    "pychalk>=2.0.1",
    "python-dotenv>=1.2.2",
    "selenium>=4.41.0",
    "validators>=0.35.0",
]
